backend/requirements.txt
appdirs==1.4.3
asgiref==3.2.10
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.0
distro==1.4.0
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
drf-yasg==1.17.1
html5lib==1.0.1
idna==2.10
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.2
lockfile==0.12.2
Markdown==3.3
MarkupSafe==1.1.1
msgpack==0.6.2
mysqlclient==2.0.1
openapi-codec==1.3.2
packaging==20.4
pep517==0.8.2
Pillow==7.2.0
progress==1.5
pyparsing==2.4.7
python-decouple==3.3
pytoml==0.1.21
pytz==2020.1
requests==2.24.0
retrying==1.3.3
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.0
uritemplate==3.0.1
urllib3==1.25.10
webencodings==0.5.1
